GNL con errores pendientes mallocs
GNL con errores pendientes

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: rbuitrag <rbuitrag@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/29 13:12:44 by rbuitrag          #+#    #+#             */
-/*   Updated: 2024/04/05 13:49:24 by rbuitrag         ###   ########.fr       */
+/*   Updated: 2024/04/08 16:41:36 by rbuitrag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 # endif
 
 char	*get_next_line(int fd);
-int		ft_strlen(char *s);
+size_t	ft_strlen(char *s);
 char	*ft_strchr(char *s, int c);
 void	*ft_calloc(size_t count, size_t size);
 char	*ft_strjoin(char *s1, char *s2);
